remove broken vasconcellos 3dof version
