+++ data/3dof.py
@@ -3,188 +3,329 @@ import plotly.graph_objects as go
 from plotly.subplots import make_subplots
 from dataclasses import dataclass
 import scipy as sp
+from scipy.integrate import solve_ivp
 
 @dataclass
 class Vars:
-    b: float # half chord length
-    a: float # nd elastic axis position relative to semi chord
-    c: float
-    rho: float # fluid density (kg/m^3)
-    m_w: float # mass of the wing (kg)
-    m_t: float # mass of wing + flap (kg)
-    zeta_alpha: float # wing pitch damping ratio
-    zeta_y: float # wing heave damping ratio
-    zeta_beta: float # flap pitch damping ratio
-    k_alpha: float # wing pitch stiffness (N/m)
-    k_y: float # wing heave stiffness (N/m)
-    k_beta: float # flap pitch stiffness (N/m)
-    x_alpha: float # distance between wing cg relative to elastic axis
-    x_beta: float # distance between flap cg relative to elastic axis
-    r_alpha: float # wing radius of gyration
-    r_beta: float # flap radius of gyration
+    a: float  # Dimensionless distance between mid-chord and EA (-0.5)
+    b: float  # Semi-chord (0.127 m)
+    c: float  # Dimensionless distance between flap hinge and mid-chord (0.5); Plunge structural damping coefficient per unit span (1.7628 kg/ms)
+    I_alpha: float  # Mass moment of inertia of the wing-flap about wing EA per unit span (0.01347 kgm)
+    I_beta: float  # Mass moment of inertia of the flap about the flap hinge line per unit span (0.0003264 kgm)
+    k_h: float # linear strctural stiffness coefficient of plunging per unit span
+    k_alpha: float  # Linear structural stiffness coefficient of plunging per unit span (2818.8 kg/ms²)
+    k_beta: float  # Linear structural stiffness coefficient of pitching per unit span (37.34 kgm/s²); Linear structural stiffness coefficient of flap per unit span (3.9 kgm/s²); Mass of wing-aileron per span (1.558 kg/m)
+    m: float  # Mass of wing-aileron per span
+    m_t: float  # Mass of wing-aileron and the supports per span
+    r_alpha: float # dimensionless radius of gyration around elastic axis
+    r_beta: float # dimensionless radius of gyration around flap hinge axis
+    S_alpha: float # static mass moment of wing-flap about wing EA per unit span
+    S_beta: float # static mass moment of flap about flap hinge line per unit span
+    x_alpha: float # dimensionless distance between airfoil EA and the center of gravity
+    x_beta: float # dimensionless distance between flap center of gravity and flap hinge axis
+    omega_h: float # uncoupled plunge natural frequency
+    omega_alpha: float # uncoupled pitch natural frequency
+    omega_beta: float # uncoupled flap natural frequency
+    rho: float # fluid density
+    zeta_h: float # plunge damping ratio
+    zeta_alpha: float # pitch damping ratio
+    zeta_beta: float # flap damping ratio
     U: float # velocity
 
-def newmark_beta_step(M, C, K, u_i, v_i, a_i, delta_F, dt, beta=1/4, gamma=1/2):
-    """
-    Implicit Newmark-Beta Method for Structural Dynamics.
+def alpha_freeplay(alpha, M0 = 0.0, Mf = 0.0, delta = np.radians(4.24), a_f = np.radians(-2.12)):
+    if (alpha < a_f):
+        return M0 + alpha - a_f
+    elif (alpha >= a_f and alpha <= (a_f + delta)):
+        return M0 + Mf * (alpha - a_f)
+    else: # alpha > a_F + delta
+        return M0 + alpha - a_f + delta * (Mf - 1)
 
-    Parameters:
-    - M, C, K: Mass, Damping, and Stiffness matrices (n x n).
-    - beta, gamma: Newmark parameters.
+def alpha_linear(alpha):
+    return alpha
 
-    Returns: incremental variation for displacement, velocity, and acceleration.
-    """
+def create_monolithic_system(y0: np.ndarray, v: Vars):
+    # def monolithic_system(t, y: np.ndarray):
+    #     A = np.zeros((8,8))
 
-    # Precompute constants
-    x2 = 1
-    x1 = gamma / (beta * dt)
-    x0 = 1 / (beta * dt**2)
-    xd0 = 1 / (beta * dt)
-    xd1 = gamma / beta
-    xdd0 = 1/(2*beta)
-    xdd1 = - dt * (1 - gamma / (2*beta))
+    #     M_s = np.zeros((3,3)) # dimensionless structural intertia matrix
+    #     D_s = np.zeros((3,3)) # dimensionless structural damping matrix
+    #     K_s = np.zeros((3,3)) # dimensionless structural stiffness matrix
 
-    # Effective stiffness matrix
-    K_eff = x0 * M + x1 * C + x2 * K
-    F_eff = delta_F + M @ (xd0 * v_i + xdd0 * a_i) + C @ (xd1 * v_i + xdd1 * a_i)
-    du = np.linalg.solve(K_eff, F_eff)
-    dv = x1 * du - xd1 * v_i - xdd1 * a_i
-    da = x0 * du - xd0 * v_i - xdd0 * a_i
+    #     sigma = v.omega_h / v.omega_alpha
+    #     V = v.U / (v.b * v.omega_alpha)
+    #     mu = v.m / (np.pi * v.rho * v.b**2)
 
-    return du, dv, da
+    #     M_s[0, 0] = v.m_t / v.m
+    #     M_s[0, 1] = v.x_alpha
+    #     M_s[0, 2] = v.x_beta
+    #     M_s[1, 0] = v.x_alpha
+    #     M_s[1, 1] = v.r_alpha**2
+    #     M_s[1, 2] = (v.c - v.a) * v.x_beta + v.r_beta**2
+    #     M_s[2, 0] = v.x_beta
+    #     M_s[2, 1] = (v.c - v.a) * v.x_beta + v.r_beta**2
+    #     M_s[2, 2] = v.r_beta**2
+    #     M_s = mu * M_s
 
-def monolithic_system_3dof(v: Vars):
-    """
-    Creates the M, C, K matrices of the coupled 3dof aeroelastic system
-    q = [y, alpha, beta, x] where x is the augmented variable    
-    $$
-    M =\left[\begin{array}{cccc}
-    m_T+\pi \rho b^2 & S_\alpha-a \pi \rho b^3 & S_\beta-\rho b^3 T_1 & 0 \\
-    S_\alpha-a \pi \rho b^3 & I_\alpha+\pi\left(\frac{1}{8}+a^2\right) \rho b^4 & I_\beta+S_\beta(c-a) b-\rho b^4\left[T_7+(c-a) T_1\right] & 0 \\
-    S_\beta-\rho b^3 T_1 & I_\beta+S_\beta(c-a) b+2 \rho b^4 T_{13} & I_\beta-\frac{\rho b^4}{\pi} T_3 & 0 \\
-    0 & 0 & 0 & 1
-    \end{array}\right]
-    $$
-    $$
-    C =\left[\begin{array}{ccc}
-    d_y+2 \pi \rho b U\left(c_0-c_1-c_3\right) & \left(1+\left(c_0-c_1-c_3\right)(1-2 a)\right) \pi \rho b^2 U & \left(T_{11}\left(c_0-c_1-c_3\right)-T_4\right) \rho U b^2 & 2 \pi \rho U^2 b\left(c_1 c_2+c_3 c_4\right)\\
-    -2 \pi \rho b^2 U\left(a+\frac{1}{2}\right)\left(c_0-c_1-c_3\right) & d_\alpha+\left(\frac{1}{2}-a\right)\left(1-\left(c_0-c_1-c_3\right)(1+2 a)\right) \pi \rho b^3 U & {\left[T_1-T_8-(c-a) T_4+\frac{T_{11}}{2}-(2 a+1) T_{11}\left(c_0-c_1-c_3\right)\right] \rho b^3 U} & {-2 \pi \rho U^2 b^2\left(a+\frac{1}{2}\right)\left(c_1 c_2+c_3 c_4\right)} \\
-    \rho U b^2 T_{12}\left(c_0-c_1-c_3\right) & {\left[T_4\left(a-\frac{1}{2}\right)-T_1-2 T_9+T_{12}\left(\frac{1}{2}-a\right)\left(c_0-c_1-c_3\right)\right] \rho U b^3} & d_\beta+\left[T_{11} T_{12}\left(c_0-c_1-c_3\right)-T_4 T_{11}\right] \frac{\rho U b^3}{2 \pi} & {\rho U^2 b^2 T_{12}\left(c_1 c_2+c_3 c_4\right)} \\
-    -\frac{1}{b} & a-\frac{1}{2} & {-\frac{T_{11}}{2 \pi}} & {\left(c_2+c_4\right) \frac{U}{b}}
-    \end{array}\right]
-    $$
-    $$
-    K =\left[\begin{array}{cccc}
-    k_{y 0} & 2 \pi \rho U^2 b\left(c_0-c_1-c_3\right) & 2 \rho U^2 b T_{10}\left(c_0-c_1-c_3\right) & 2 \pi \rho U^3\left(c_2 c_4\left(c_1+c_3\right)\right) \\
-    0 & k_{\alpha 0}-2 \pi\left(\frac{1}{2}+a\right)\left(c_0-c_1-c_3\right) \rho b^2 U^2 & \rho U^2 b^2\left[T_4+T_{10}+T_{10}(2 a+1)\left(c_0-c_1-c_3\right)\right] & -2 \pi \rho U^3 b\left(a+\frac{1}{2}\right)\left(c_2 c_4\left(c_1+c_3\right)\right) \\
-    0 & \rho U^2 b^2 T_{12}\left(c_0-c_1-c_3\right) & k_{\beta 0}+\left(T_5-T_4 T_{10}+T_{10} T_{12}\left(c_0-c_1-c_3\right)\right) \frac{\rho U^2 b^2}{\pi} & \rho U^3 b T_{12}\left(c_2 c_4\left(c_1+c_3\right)\right) \\
-    0 & -\frac{U}{b} & -\frac{T_{10} U}{b \pi} & c_2 c_4 \frac{U^2}{b^2}
-    \end{array}\right]
-    $$
+    #     D_s[0, 0] = sigma * v.zeta_h
+    #     D_s[1, 1] = v.r_alpha**2 * v.zeta_alpha
+    #     D_s[2, 2] = (v.omega_beta / v.omega_alpha) * v.r_beta**2 * v.zeta_beta
+    #     D_s = 2 * mu * D_s
 
+    #     gamma = 10 # Cubic factor (0 for linear spring)
+    #     K_s[0, 0] = sigma**2 * (1 + gamma * y[3]**2)
+    #     K_s[1, 1] = v.r_alpha**2
+    #     K_s[2, 2] = (v.omega_beta / v.omega_alpha)**2 * v.r_beta**2
+    #     K_s = mu * K_s
 
-    """
-    S_alpha = v.x_alpha * v.m_w * v.b
-    S_beta = v.x_beta * v.m_w * v.b
-    I_alpha = v.r_alpha**2 * v.m_w * v.b ** 2
-    I_beta = v.r_beta**2 * v.m_w * v.b ** 2
+    #     sqrt_1_minus_c2 = np.sqrt(1 - v.c**2)
+    #     acos_c = np.arccos(v.c)
 
-    omega_alpha = np.sqrt(v.k_alpha / (v.m_w * (v.r_alpha * v.b)**2))
-    omega_beta = np.sqrt(v.k_beta / (v.m_w * (v.r_beta * v.b)**2))
-    omega_y = np.sqrt(v.k_y / v.m_w)
-    d_y = 2*v.m_w*omega_y*v.zeta_y
-    d_alpha = 2*I_alpha*omega_alpha*v.zeta_alpha
-    d_beta = 2*I_beta*omega_beta*v.zeta_beta
+    #     T1 = (-1/3) * sqrt_1_minus_c2 * (2 + v.c**2) + v.c * acos_c
+    #     T2 = v.c * (1 - v.c**2) - sqrt_1_minus_c2 * (1 + v.c**2) * acos_c + v.c * (acos_c)**2
+    #     T3 = -((1/8) + v.c**2) * (acos_c)**2 + (1/4) * v.c * sqrt_1_minus_c2 * acos_c * (7 + 2*v.c**2) - (1/8) * (1 - v.c**2) * (5 * v.c**2 + 4)
+    #     T4 = -acos_c + v.c * sqrt_1_minus_c2
+    #     T5 = -(1 - v.c**2) - (acos_c)**2 + 2 * v.c * sqrt_1_minus_c2 * acos_c
+    #     T6 = T2
+    #     T7 = -((1/8) + v.c**2) * acos_c + (1/8) * v.c * sqrt_1_minus_c2 * (7 + 2 * v.c**2)
+    #     T8 = (-1/3) * sqrt_1_minus_c2 * (2 * v.c**2 + 1) + v.c * acos_c
+    #     T9 = (1/2) * ((1/3) * (sqrt_1_minus_c2)**3 + v.a * T4)
+    #     T10 = sqrt_1_minus_c2 + acos_c
+    #     T11 = acos_c * (1 - 2 * v.c) + sqrt_1_minus_c2 * (2 - v.c)
+    #     T12 = sqrt_1_minus_c2 * (2 + v.c) - acos_c * (2 * v.c + 1)
+    #     T13 = (1/2) * (-T7 - (v.c - v.a) * T1)
+    #     T14 = (1/16) + (1/2) * v.a * v.c
 
-    sqrt_1_minus_c2 = np.sqrt(1 - v.c**2)
-    acos_c = np.arccos(v.c)
+    #     M_a = np.zeros((3,3)) # dimensionless aerodynamic inertia matrix
+    #     D_a = np.zeros((3,3)) # dimensionless aerodynamic damping matrix
+    #     K_a = np.zeros((3,3)) # dimensionless aerodynamic stiffness matrix
+    #     L_delta = np.zeros((3, 2)) # dimensionless aero lagging matrix
+    #     Q_a = np.zeros((2, 3)) # matrix of terms multiplied by the modal accelerations
+    #     Q_v = np.zeros((2, 3)) # matrix of terms multiplied by the modal velocities
+    #     L_lambda = np.zeros((2, 2))
 
-    T1 = (-1/3) * sqrt_1_minus_c2 * (2 + v.c**2) + v.c * acos_c
-    T2 = v.c * (1 - v.c**2) - sqrt_1_minus_c2 * (1 + v.c**2) * acos_c + v.c * (acos_c)**2
-    T3 = -((1/8) + v.c**2) * (acos_c)**2 + (1/4) * v.c * sqrt_1_minus_c2 * acos_c * (7 + 2*v.c**2) - (1/8) * (1 - v.c**2) * (5 * v.c**2 + 4)
-    T4 = -acos_c + v.c * sqrt_1_minus_c2
-    T5 = -(1 - v.c**2) - (acos_c)**2 + 2 * v.c * sqrt_1_minus_c2 * acos_c
-    T6 = T2
-    T7 = -((1/8) + v.c**2) * acos_c + (1/8) * v.c * sqrt_1_minus_c2 * (7 + 2 * v.c**2)
-    T8 = (-1/3) * sqrt_1_minus_c2 * (2 * v.c**2 + 1) + v.c * acos_c
-    T9 = (1/2) * ((1/3) * (sqrt_1_minus_c2)**3 + v.a * T4)
-    T10 = sqrt_1_minus_c2 + acos_c
-    T11 = acos_c * (1 - 2 * v.c) + sqrt_1_minus_c2 * (2 - v.c)
-    T12 = sqrt_1_minus_c2 * (2 + v.c) - acos_c * (2 * v.c + 1)
-    T13 = (1/2) * (-T7 - (v.c - v.a) * T1)
-    T14 = (1/16) + (1/2) * v.a * v.c
+    #     M_a[0, 0] = -1
+    #     M_a[0, 1] = v.a
+    #     M_a[0, 2] = T1 / np.pi
+    #     M_a[1, 0] = v.a
+    #     M_a[1, 1] = -(1/8 + v.a**2)
+    #     M_a[1, 2] = -2 * T13 / np.pi
+    #     M_a[2, 0] = T1 / np.pi
+    #     M_a[2, 1] = -2 * T13 / np.pi
+    #     M_a[2, 2] = T3 / (np.pi**2)
 
-    c0 = 1.0
-    c1 = 0.165
-    c2 = 0.0455
-    c3 = 0.335
-    c4 = 0.3
+    #     D_a[0, 0] = (-2)
+    #     D_a[0, 1] = (-2 * (1 - v.a))
+    #     D_a[0, 2] = (T4 - T11) / np.pi
+    #     D_a[1, 0] = (1 + 2 * v.a)
+    #     D_a[1, 1] = (v.a * (1 - 2 * v.a))
+    #     D_a[1, 2] = (T8 - T1 + (v.c - v.a) * T4 + v.a * T11) / np.pi
+    #     D_a[2, 0] = (-T12 / np.pi)
+    #     D_a[2, 1] = (2 * T9 + T1 + (T12 - T4) * (v.a - 0.5)) / np.pi
+    #     D_a[2, 2] = (T11 * (T4 - T12)) / (2 * np.pi**2)
+    #     D_a = V * D_a
 
-    M = np.zeros((4,4), dtype=np.float64)
-    C = np.zeros((4,4), dtype=np.float64)
-    K = np.zeros((4,4), dtype=np.float64)
+    #     K_a[0, 0] = 0
+    #     K_a[0, 1] = (-2)
+    #     K_a[0, 2] = (-2 * T10) / np.pi
+    #     K_a[1, 0] = 0
+    #     K_a[1, 1] = (1 + 2 * v.a)
+    #     K_a[1, 2] = (2 * v.a * T10 - T4) / np.pi
+    #     K_a[2, 0] = 0
+    #     K_a[2, 1] = (-T12) / np.pi
+    #     K_a[2, 2] = (-1 / np.pi**2) * (T5 - T10 * (T4 - T12))
+    #     K_a = V**2 * K_a
+        
+    #     # Difference lies between coeffs of R.T Jones and W.P Jones
+    #     # Yung p220
+    #     d1 = 0.165
+    #     d2 = 0.335
+    #     # l1 = 0.041
+    #     # l2 = 0.320
+    #     l1 = 0.0455
+    #     l2 = 0.300
 
-    M[0, 0] = v.m_t + np.pi * v.rho * v.b**2
-    M[0, 1] = S_alpha - v.a * np.pi * v.rho * v.b**3
-    M[0, 2] = S_beta - v.rho * v.b**3 * T1
-    M[0, 3] = 0
+    #     L_delta[0, 0] = d1
+    #     L_delta[0, 1] = d2
+    #     L_delta[1, 0] = - (0.5 + v.a) * d1
+    #     L_delta[1, 1] = - (0.5 + v.a) * d2
+    #     L_delta[2, 0] = T12 * d1 / (2 * np.pi)
+    #     L_delta[2, 1] = T12 * d2 / (2 * np.pi)
+    #     L_delta = 2 * V * L_delta
 
-    M[1, 0] = S_alpha - v.a * np.pi * v.rho * v.b**3
-    M[1, 1] = I_alpha + np.pi * (1/8 + v.a**2) * v.rho * v.b**4
-    M[1, 2] = I_beta + S_beta * (v.c - v.a) * v.b - v.rho * v.b**4 * (T7 + (v.c - v.a) * T1)
-    M[1, 3] = 0
+    #     Q_a[0, 0] = 1
+    #     Q_a[0, 1] = 0.5 - v.a
+    #     Q_a[0, 2] = T11 / (2 * np.pi)
+    #     Q_a[1, 0] = 1
+    #     Q_a[1, 1] = 0.5 - v.a
+    #     Q_a[1, 2] = T11 / (2 * np.pi)
 
-    M[2, 0] = S_beta - v.rho * v.b**3 * T1
-    M[2, 1] = I_beta + S_beta * (v.c - v.a) * v.b + 2 * v.rho * v.b**4 * T13
-    M[2, 2] = I_beta - (v.rho * v.b**4 / np.pi) * T3
-    M[2, 3] = 0
+    #     Q_v[0, 1] = 1
+    #     Q_v[0, 2] = T10 / np.pi
+    #     Q_v[1, 1] = 1
+    #     Q_v[1, 2] = T10 / np.pi
+    #     Q_v = V * Q_v
 
-    M[3, 0] = 0
-    M[3, 1] = 0
-    M[3, 2] = 0
-    M[3, 3] = 1
+    #     L_lambda[0, 0] = - l1
+    #     L_lambda[1, 1] = - l2
+    #     L_lambda = V * L_lambda
 
-    C[0, 0] = d_y + 2 * np.pi * v.rho * v.b * v.U * (c0 - c1 - c3)
-    C[0, 1] = (1 + (c0 - c1 - c3) * (1 - 2 * v.a)) * np.pi * v.rho * v.b**2 * v.U
-    C[0, 2] = (T11 * (c0 - c1 - c3) - T4) * v.rho * v.U * v.b**2
-    C[0, 3] = 2 * np.pi * v.rho * v.U**2 * v.b * (c1 * c2 + c3 * c4)
+    #     inv_inert_mat = np.linalg.inv(M_s - M_a)
+    #     A[0:3, 0:3] = - inv_inert_mat @ (D_s - D_a)
+    #     A[0:3, 3:6] = - inv_inert_mat @ (K_s - K_a)
+    #     A[0:3, 6:8] = inv_inert_mat @ L_delta
+    #     A[3:6, 0:3] = np.eye(3)
+    #     A[6:8, 0:3] = Q_a @ A[0:3, 0:3]+ Q_v
+    #     A[6:8, 3:6] = Q_a @ A[0:3, 3:6]
+    #     A[6:8, 6:8] = Q_a @ A[0:3, 6:8] + L_lambda
 
-    C[1, 0] = -2 * np.pi * v.rho * v.b**2 * v.U * (v.a + 0.5) * (c0 - c1 - c3)
-    C[1, 1] = d_alpha + (0.5 - v.a) * (1 - (c0 - c1 - c3) * (1 + 2 * v.a)) * np.pi * v.rho * v.b**3 * v.U
-    C[1, 2] = (T1 - T8 - (v.c - v.a) * T4 + T11 / 2 - (2 * v.a + 1) * T11 * (c0 - c1 - c3)) * v.rho * v.b**3 * v.U
-    C[1, 3] = -2 * np.pi * v.rho * v.U**2 * v.b**2 * (v.a + 0.5) * (c1 * c2 + c3 * c4)
+    #     return A @ y
 
-    C[2, 0] = v.rho * v.U * v.b**2 * T12 * (c0 - c1 - c3)
-    C[2, 1] = (T4 * (v.a - 0.5) - T1 - 2 * T9 + T12 * (0.5 - v.a) * (c0 - c1 - c3)) * v.rho * v.U * v.b**3
-    C[2, 2] = d_beta + (T11 * T12 * (c0 - c1 - c3) - T4 * T11) * (v.rho * v.U * v.b**3) / (2 * np.pi)
-    C[2, 3] = v.rho * v.U**2 * v.b**2 * T12 * (c1 * c2 + c3 * c4)
+    def monolithic_system(t, y: np.ndarray):
+        M1 = np.zeros((8, 8))
+        M2 = np.zeros((8, 8))
+        yn = np.zeros(8) # nonlinear terms
 
-    C[3, 0] = -1 / v.b
-    C[3, 1] = v.a - 0.5
-    C[3, 2] = -T11 / (2 * np.pi)
-    C[3, 3] = (c2 + c4) * v.U / v.b
+        M_s = np.zeros((3,3)) # dimensionless structural intertia matrix
+        D_s = np.zeros((3,3)) # dimensionless structural damping matrix
+        K_s = np.zeros((3,3)) # dimensionless structural stiffness matrix
 
-    K[0, 0] = v.k_y
-    K[0, 1] = 2 * np.pi * v.rho * v.U**2 * v.b * (c0 - c1 - c3)
-    K[0, 2] = 2 * v.rho * v.U**2 * v.b * T10 * (c0 - c1 - c3)
-    K[0, 3] = 2 * np.pi * v.rho * v.U**3 * (c2 * c4 * (c1 + c3))
+        sigma = v.omega_h / v.omega_alpha
+        V = v.U / (v.b * v.omega_alpha)
+        mu = v.m / (np.pi * v.rho * v.b**2)
 
-    K[1, 0] = 0
-    K[1, 1] = v.k_alpha - 2 * np.pi * (0.5 + v.a) * (c0 - c1 - c3) * v.rho * v.b**2 * v.U**2
-    K[1, 2] = v.rho * v.U**2 * v.b**2 * (T4 + T10 + T10 * (2 * v.a + 1) * (c0 - c1 - c3))
-    K[1, 3] = -2 * np.pi * v.rho * v.U**3 * v.b * (v.a + 0.5) * (c2 * c4 * (c1 + c3))
+        M_s[0, 0] = v.m_t / v.m
+        M_s[0, 1] = v.x_alpha
+        M_s[0, 2] = v.x_beta
+        M_s[1, 0] = v.x_alpha
+        M_s[1, 1] = v.r_alpha**2
+        M_s[1, 2] = (v.c - v.a) * v.x_beta + v.r_beta**2
+        M_s[2, 0] = v.x_beta
+        M_s[2, 1] = (v.c - v.a) * v.x_beta + v.r_beta**2
+        M_s[2, 2] = v.r_beta**2
+        M_s = mu * M_s
 
-    K[2, 0] = 0
-    K[2, 1] = v.rho * v.U**2 * v.b**2 * T12 * (c0 - c1 - c3)
-    K[2, 2] = v.k_beta + (T5 - T4 * T10 + T10 * T12 * (c0 - c1 - c3)) * (v.rho * v.U**2 * v.b**2) / np.pi
-    K[2, 3] = v.rho * v.U**3 * v.b * T12 * (c2 * c4 * (c1 + c3))
+        D_s[0, 0] = sigma * v.zeta_h
+        D_s[1, 1] = v.r_alpha**2 * v.zeta_alpha
+        D_s[2, 2] = (v.omega_beta / v.omega_alpha) * v.r_beta**2 * v.zeta_beta
+        D_s = 2 * mu * D_s
 
-    K[3, 0] = 0
-    K[3, 1] = -v.U / v.b
-    K[3, 2] = -T10 * v.U / (v.b * np.pi)
-    K[3, 3] = c2 * c4 * v.U**2 / v.b**2
+        # K_s[0, 0] = sigma**2 * (1 + gamma * y[3]**2)
+        K_s[1, 1] = v.r_alpha**2
+        # K_s[2, 2] = (v.omega_beta / v.omega_alpha)**2 * v.r_beta**2
+        K_s = mu * K_s
 
-    return M, C, K
+        sqrt_1_minus_c2 = np.sqrt(1 - v.c**2)
+        acos_c = np.arccos(v.c)
+
+        T1 = (-1/3) * sqrt_1_minus_c2 * (2 + v.c**2) + v.c * acos_c
+        T2 = v.c * (1 - v.c**2) - sqrt_1_minus_c2 * (1 + v.c**2) * acos_c + v.c * (acos_c)**2
+        T3 = -((1/8) + v.c**2) * (acos_c)**2 + (1/4) * v.c * sqrt_1_minus_c2 * acos_c * (7 + 2*v.c**2) - (1/8) * (1 - v.c**2) * (5 * v.c**2 + 4)
+        T4 = -acos_c + v.c * sqrt_1_minus_c2
+        T5 = -(1 - v.c**2) - (acos_c)**2 + 2 * v.c * sqrt_1_minus_c2 * acos_c
+        T6 = T2
+        T7 = -((1/8) + v.c**2) * acos_c + (1/8) * v.c * sqrt_1_minus_c2 * (7 + 2 * v.c**2)
+        T8 = (-1/3) * sqrt_1_minus_c2 * (2 * v.c**2 + 1) + v.c * acos_c
+        T9 = (1/2) * ((1/3) * (sqrt_1_minus_c2)**3 + v.a * T4)
+        T10 = sqrt_1_minus_c2 + acos_c
+        T11 = acos_c * (1 - 2 * v.c) + sqrt_1_minus_c2 * (2 - v.c)
+        T12 = sqrt_1_minus_c2 * (2 + v.c) - acos_c * (2 * v.c + 1)
+        T13 = (1/2) * (-T7 - (v.c - v.a) * T1)
+        T14 = (1/16) + (1/2) * v.a * v.c
+
+        M_a = np.zeros((3,3)) # dimensionless aerodynamic inertia matrix
+        D_a = np.zeros((3,3)) # dimensionless aerodynamic damping matrix
+        K_a = np.zeros((3,3)) # dimensionless aerodynamic stiffness matrix
+        L_delta = np.zeros((3, 2)) # dimensionless aero lagging matrix
+        Q_a = np.zeros((2, 3)) # matrix of terms multiplied by the modal accelerations
+        Q_v = np.zeros((2, 3)) # matrix of terms multiplied by the modal velocities
+        L_lambda = np.zeros((2, 2))
+
+        M_a[0, 0] = -1
+        M_a[0, 1] = v.a
+        M_a[0, 2] = T1 / np.pi
+        M_a[1, 0] = v.a
+        M_a[1, 1] = -(1/8 + v.a**2)
+        M_a[1, 2] = -2 * T13 / np.pi
+        M_a[2, 0] = T1 / np.pi
+        M_a[2, 1] = -2 * T13 / np.pi
+        M_a[2, 2] = T3 / (np.pi**2)
+
+        D_a[0, 0] = (-2)
+        D_a[0, 1] = (-2 * (1 - v.a))
+        D_a[0, 2] = (T4 - T11) / np.pi
+        D_a[1, 0] = (1 + 2 * v.a)
+        D_a[1, 1] = (v.a * (1 - 2 * v.a))
+        D_a[1, 2] = (T8 - T1 + (v.c - v.a) * T4 + v.a * T11) / np.pi
+        D_a[2, 0] = (-T12 / np.pi)
+        D_a[2, 1] = (2 * T9 + T1 + (T12 - T4) * (v.a - 0.5)) / np.pi
+        D_a[2, 2] = (T11 * (T4 - T12)) / (2 * np.pi**2)
+        D_a = V * D_a
+
+        K_a[0, 0] = 0
+        K_a[0, 1] = (-2)
+        K_a[0, 2] = (-2 * T10) / np.pi
+        K_a[1, 0] = 0
+        K_a[1, 1] = (1 + 2 * v.a)
+        K_a[1, 2] = (2 * v.a * T10 - T4) / np.pi
+        K_a[2, 0] = 0
+        K_a[2, 1] = (-T12) / np.pi
+        K_a[2, 2] = (-1 / np.pi**2) * (T5 - T10 * (T4 - T12))
+        K_a = V**2 * K_a
+        
+        # Difference lies between coeffs of R.T Jones and W.P Jones
+        # Yung p220
+        d1 = 0.165
+        d2 = 0.335
+        # l1 = 0.041
+        # l2 = 0.320
+        l1 = 0.0455
+        l2 = 0.300
+
+        L_delta[0, 0] = d1
+        L_delta[0, 1] = d2
+        L_delta[1, 0] = - (0.5 + v.a) * d1
+        L_delta[1, 1] = - (0.5 + v.a) * d2
+        L_delta[2, 0] = T12 * d1 / (2 * np.pi)
+        L_delta[2, 1] = T12 * d2 / (2 * np.pi)
+        L_delta = 2 * V * L_delta
+
+        Q_a[0, 0] = 1
+        Q_a[0, 1] = 0.5 - v.a
+        Q_a[0, 2] = T11 / (2 * np.pi)
+        Q_a[1, 0] = 1
+        Q_a[1, 1] = 0.5 - v.a
+        Q_a[1, 2] = T11 / (2 * np.pi)
+
+        Q_v[0, 1] = 1
+        Q_v[0, 2] = T10 / np.pi
+        Q_v[1, 1] = 1
+        Q_v[1, 2] = T10 / np.pi
+        Q_v = V * Q_v
+
+        L_lambda[0, 0] = - l1
+        L_lambda[1, 1] = - l2
+        L_lambda = V * L_lambda
+
+        M2[0:3, 0:3] = M_s - M_a
+        M2[3:6, 3:6] = np.eye(3)
+        M2[6:8, 6:8] = np.eye(2)
+        # M2[3:8, 3:8] = np.eye(5)
+        M2[6:8, 0:3] = - Q_a
+
+        M1[0:3, 0:3] = D_a - D_s
+        M1[0:3, 3:6] = K_a - K_s
+        M1[0:3, 6:8] = L_delta
+        M1[3:6, 0:3] = np.eye(3)
+        M1[6:8, 0:3] = Q_v
+        M1[6:8, 6:8] = L_lambda
+
+        # Nonlinear term
+        gamma = 0 # Cubic factor (0 for linear spring)
+        yn[0] = - mu * y[3] * sigma**2 * (1 + gamma * y[3]**2)
+        yn[2] = - mu * ((v.omega_beta / v.omega_alpha)**2 * v.r_beta**2) * alpha_freeplay(y[5])
+        return np.linalg.solve(M2, M1 @ y + yn)
+
+    return monolithic_system
 
 def compute_psd(t, data):
     """Compute PSD with consistent parameters"""
@@ -197,8 +338,8 @@ def compute_psd(t, data):
     mask = frequencies < 1.0
     psd_db = 10 * np.log10(psd)
 
-    # return frequencies[mask], psd_db[mask]
-    return frequencies, psd
+    return frequencies[mask], psd[mask]
+    # return frequencies, psd
 
 def add_data_and_psd(fig, time, data, name, row_data, col_data, mode='lines', marker_size=4):
     """Add time series and PSD data to plotly figure"""
@@ -251,50 +392,96 @@ def format_subplot(fig, row, col, xlabel, ylabel):
     )
 
 if __name__ == "__main__":
+    # Darabseh 2022
+    # v = Vars(
+    #     a = -0.5,
+    #     b = 0.127, # m
+    #     c = 0.5,
+    #     I_alpha = 0.01347, # kgm
+    #     I_beta = 0.0003264, # kgm
+    #     k_h = 2818.8, # kg/ms^2
+    #     k_alpha = 37.34, # kg/ms^2
+    #     k_beta = 3.9, # kg/ms^2
+    #     m = 1.558, # kg/m
+    #     m_t = 3.3843, # kg/m
+    #     r_alpha = 0.7321,
+    #     r_beta = 0.1140,
+    #     S_alpha = 0.08587, # kg
+    #     S_beta = 0.00395, # kg
+    #     x_alpha = 0.4340,
+    #     x_beta = 0.02,
+    #     omega_h = 42.5352, # Hz
+    #     omega_alpha = 52.6506, # Hz
+    #     omega_beta = 109.3093,
+    #     rho = 1.225, # kg/m^3
+    #     zeta_h = 0.0133,
+    #     zeta_alpha = 0.01626,
+    #     zeta_beta = 0.0133,
+    #     U = 23.72 # m/s
+    # )
+    
+    # Conner
     v = Vars(
-        b = 0.127,
         a = -0.5,
+        b = 0.127, # m
         c = 0.5,
-        rho = 1.225,
-        m_w = 1.7025,
-        m_t = 3.5021,
-        zeta_alpha = 0.01,
-        zeta_y = 0.012,
-        zeta_beta = 0.015,
-        k_alpha = 1114.0,
-        k_y = 17240,
-        k_beta = 73.07,
-        x_alpha = 0.3294,
-        x_beta = 0.01795,
-        r_alpha = 0.6840,
-        r_beta = 0.07336,
-        U = 24.0
+        I_alpha = 0.01347, # kgm
+        I_beta = 0.0003264, # kgm
+        k_h = 2818.8, # kg/ms^2
+        k_alpha = 37.34, # kg/ms^2
+        k_beta = 3.9, # kg/ms^2
+        m = 1.5666, # kg/m
+        m_t = 3.39298, # kg/m
+        r_alpha = 0.7321,
+        r_beta = 0.1140,
+        S_alpha = 0.08587, # kg
+        S_beta = 0.00395, # kg
+        x_alpha = 0.4340,
+        x_beta = 0.02,
+        omega_h = 42.5352, # rad/s
+        omega_alpha = 52.6506, # rad/s
+        omega_beta = 109.3093, # rad/s
+        rho = 1.225, # kg/m^3
+        zeta_h = 0.0113,
+        zeta_alpha = 0.01626,
+        zeta_beta = 0.0115,
+        U = 0.49 * 23.9 # m/s
     )
 
-    dt = 0.01
-    t_final = 5.0
+    dt = 0.02
+    t_final = 10 * v.omega_alpha
     vec_t = np.arange(0, t_final, dt)
     n = len(vec_t)
 
-    M, C, K = monolithic_system_3dof(v)
+    # hd, ad, bd, h, a, b, x1, x2
+    y0 = np.array([
+        0, # dh / dt
+        0, # dalpha / dt
+        0, # dbeta / dt
+        0.01 / v.b, # h
+        0, # alpha
+        0, # beta
+        0, # x1
+        0  # x2
+    ], dtype=np.float64)
 
-    # Historical data
-    u = np.zeros((4, n), dtype=np.float64)
-    v = np.zeros((4, n), dtype=np.float64)
-    a = np.zeros((4, n), dtype=np.float64)
-    F = np.zeros((4, n), dtype=np.float64) # external forces / nonlinear terms
-    dF = np.zeros(4, dtype=np.float64)
+    # y0 = np.array([
+    #     0, # dh / dt
+    #     0, # dalpha / dt
+    #     0, # dbeta / dt
+    #     0, # h
+    #     0, # alpha
+    #     np.radians(-2.12), # beta
+    #     0, # x1
+    #     0  # x2
+    # ], dtype=np.float64)
 
-    # Initial conditions
-    u[:, 0] = np.array([0.01, 0, 0, 0], dtype=np.float64)
-    v[:, 0] = np.array([0, 0, 0, 0], dtype=np.float64)
-    a[:, 0] = np.linalg.solve(M, F[:, 0] - C @ v[:,0] - K @ u[:,0])
+    system = create_monolithic_system(y0, v)
+    mono = solve_ivp(system, (0, t_final), y0, t_eval=vec_t, method='RK45')
     
-    for i in range(n-1):
-        du, dv, da = newmark_beta_step(M, C, K, u[:,i], v[:,i], a[:,i], dF, dt)
-        u[:, i+1] = u[:, i] + du
-        v[:, i+1] = v[:, i] + dv
-        a[:, i+1] = a[:, i] + da
+    if (not mono.success):
+        print("Monolithic solver failed")
+        exit(1)
 
     # Plotting
     fig = make_subplots(
@@ -310,12 +497,12 @@ if __name__ == "__main__":
         vertical_spacing=0.1,
         horizontal_spacing=0.08
     )
-    add_data_and_psd(fig, vec_t, u[0, :], "Theodorsen", 1, 1)
-    add_data_and_psd(fig, vec_t, v[0, :], "Theodorsen", 1, 2)
-    add_data_and_psd(fig, vec_t, np.degrees(u[1, :]), "Theodorsen", 3, 1)
-    add_data_and_psd(fig, vec_t, np.degrees(v[1, :]), "Theodorsen", 3, 2)
-    add_data_and_psd(fig, vec_t, np.degrees(u[2, :]), "Theodorsen", 5, 1)
-    add_data_and_psd(fig, vec_t, np.degrees(v[2, :]), "Theodorsen", 5, 2)
+    add_data_and_psd(fig, vec_t, mono.y[3, :], "Theodorsen", 1, 1)
+    add_data_and_psd(fig, vec_t, mono.y[0, :], "Theodorsen", 1, 2)
+    add_data_and_psd(fig, vec_t, np.degrees(mono.y[4, :]), "Theodorsen", 3, 1)
+    add_data_and_psd(fig, vec_t, np.degrees(mono.y[1, :]), "Theodorsen", 3, 2)
+    add_data_and_psd(fig, vec_t, np.degrees(mono.y[5, :]), "Theodorsen", 5, 1)
+    add_data_and_psd(fig, vec_t, np.degrees(mono.y[2, :]), "Theodorsen", 5, 2)
 
     format_subplot(fig, 1, 1, "t", r"y")
     format_subplot(fig, 1, 2, "t", r"$\dot{y}$")
